React: rename identity function to noOP
