+++ client/src/utils/index.js
@@ -3,9 +3,9 @@ import auth from './auth';
 import changeCase from './change_case';
 import compareSpots from './compare-spots';
 import errorEmitter from './error-emitter';
-import identityFunction from './identity-function';
 import isEmptyObject from './is_empty_object';
+import noOp from './no_op';
 import * as routes from './routes';
 import screenSizes from './screen-sizes';
 
-export { api, auth, changeCase, compareSpots, errorEmitter, identityFunction, isEmptyObject, routes, screenSizes };
+export { api, auth, changeCase, compareSpots, errorEmitter, isEmptyObject, noOp, routes, screenSizes };
